refactor(divider): move dividerSVG logic into Divider, update readme
